Rename 'components/MonitoringCommentFrame.vue' to 'components/index/SiteTopUpper/MonitoringComment/Frame.vue'
