+++ components/index/SiteTopUpper/MonitoringCommentCard.vue
@@ -34,14 +34,14 @@
       <v-row>
         <v-col cols="12" sm="12" md="6" lg="6">
           <h4>{{ $t('感染状況') }}</h4>
-          <monitoring-comment-frame
+          <frame
             :level="monitoringItems.data['総括コメント-感染状況'].level - 1"
             :comment="monitoringItems.data['総括コメント-感染状況'].label"
           />
         </v-col>
         <v-col cols="12" sm="12" md="6" lg="6">
           <h4>{{ $t('医療提供体制') }}</h4>
-          <monitoring-comment-frame
+          <frame
             :level="monitoringItems.data['総括コメント-医療提供体制'].level - 1"
             :comment="monitoringItems.data['総括コメント-医療提供体制'].label"
           />
@@ -55,13 +55,13 @@
 import Vue from 'vue'
 
 import AppLink from '@/components/AppLink.vue'
-import MonitoringCommentFrame from '@/components/MonitoringCommentFrame.vue'
+import Frame from '@/components/index/SiteTopUpper/MonitoringComment/Frame.vue'
 import monitoringItems from '@/data/monitoring_items.json'
 
 export default Vue.extend({
   components: {
     AppLink,
-    MonitoringCommentFrame,
+    Frame,
   },
   data() {
     return {
